Task 17: add event controller (#20)
* add event controller

* refactoring

* refactoring

+++ src/main/java/com/crudshop/demo/interaction/ExchangeRateClientImplStub.java
@@ -1,11 +1,12 @@
 package com.crudshop.demo.interaction;
 
-import com.crudshop.demo.util.Currency;
-import com.crudshop.demo.util.ExchangeRate;
+import com.crudshop.demo.currency.ExchangeRate;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.context.annotation.Primary;
 import org.springframework.stereotype.Component;
+
+import java.math.BigDecimal;
 
 @Component
 @Primary
@@ -15,7 +16,7 @@ public class ExchangeRateClientImplStub implements ExchangeRateClient {
     @Override
     public ExchangeRate getExchangeRate() {
         ExchangeRate exchangeRate = new ExchangeRate();
-        exchangeRate.setExchangeRateRUB(20.0);
+        exchangeRate.setExchangeRateRUB(BigDecimal.valueOf(20.0));
         log.info("Значение получено из стаба");
 
         return exchangeRate;
